FIX: Cache breadcrumb-item template
…to avoid re-evaulation right before destroying.

With `DeferredTrackedSet` we delay both adding and removing elements from the set. That means when you're transitioning between routes, and breadcrums change, both old and new breadcrums are rendered (briefly, in a first render pass)

And since the arguments for the old breadcrums can be (and often are) destroyed - it would blow up the renderer. By caching the template it will reuse it in that first pass.